añil/U2/Akinator/pyproject.toml
[tool.poetry]
name = "akinator"
version = "0.1.0"
description = ""
authors = ["JK"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruamel-yaml = "^0.18.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
